+++ src/pages/stored-codes.tsx
@@ -1,0 +1,3 @@
+import StoredCodes from "lib/pages/stored-codes";
+
+export default StoredCodes;